separate new and old folders
