--- test/integration_rework/001_simple_copy_tests/models/advanced_incremental.sql
@@ -1,17 +1,0 @@
-{{
-  config(
-    materialized = "incremental",
-    unique_key = "id",
-    persist_docs = {"relation": true}
-  )
-}}
-
-
-select *
-from {{ ref('seed') }}
-
-{% if is_incremental() %}
-
-    where id > (select max(id) from {{this}})
-
-{% endif %}
