--- test/integration_rework/001_simple_copy_tests/models/compound_sort.sql
@@ -1,9 +1,0 @@
-{{
-  config(
-    materialized = "table",
-    sort = 'first_name',
-    sort_type = 'compound'
-  )
-}}
-
-select * from {{ ref('seed') }}
