--- test/integration_rework/001_simple_copy_tests/models/disabled.sql
@@ -1,8 +1,0 @@
-{{
-  config(
-    materialized = "view",
-    enabled = False
-  )
-}}
-
-select * from {{ ref('seed') }}
