--- test/integration_rework/001_simple_copy_tests/models/empty.sql
@@ -1,5 +1,0 @@
-
-
-
-
-
