--- test/integration_rework/001_simple_copy_tests/models/get_and_ref.sql
@@ -1,3 +1,0 @@
-{%- do adapter.get_relation(database=target.database, schema=target.schema, identifier='materialized') -%}
-
-select * from {{ ref('materialized') }}
