--- test/integration_rework/001_simple_copy_tests/models/incremental.sql
@@ -1,11 +1,0 @@
-{{
-  config(
-    materialized = "incremental"
-  )
-}}
-
-select * from {{ ref('seed') }}
-
-{% if is_incremental() %}
-    where id > (select max(id) from {{this}})
-{% endif %}
