--- test/integration_rework/001_simple_copy_tests/models/interleaved_sort.sql
@@ -1,9 +1,0 @@
-{{
-  config(
-    materialized = "table",
-    sort = ['first_name', 'last_name'],
-    sort_type = 'interleaved'
-  )
-}}
-
-select * from {{ ref('seed') }}
