--- test/integration_rework/001_simple_copy_tests/models/schema.yml
@@ -1,7 +1,0 @@
-version: 2
-models:
-- name: disabled
-  columns:
-  - name: id
-    tests:
-    - unique
