--- test/integration_rework/001_simple_copy_tests/models/view_model.sql
@@ -1,7 +1,0 @@
-{{
-  config(
-    materialized = "view"
-  )
-}}
-
-select * from {{ ref('seed') }}
