--- test/integration_rework/conftest.py
@@ -1,9 +1,0 @@
-import pytest
-
-
-def pytest_collection_modifyitems(items):
-    for item in items:
-        if "interface" in item.nodeid:
-            item.add_marker(pytest.mark.interface)
-        elif "event" in item.nodeid:
-            item.add_marker(pytest.mark.event)
